+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Bug fixes
 ---------
 
 * Eliminate unnecessary property write-back when accessed via a getter since property getters are assumed to be _pure_.
+* Writing to a property of an indexed valued obtained via an indexer now works properly by writing back the changed value via an index setter.
 
 Enhancements
 ------------